llm_backend/requirements.txt
beautifulsoup4==4.12.3
flasgger==0.9.7.1
Flask==3.1.0
flask_session==0.8.0
flask-cors==5.0.0
Flask_SocketIO==5.4.1
langchain==0.3.9
langchain_chroma==0.1.4
langchain_core==0.3.21
langchain_openai==0.2.10
langchain_text_splitters==0.3.2
openai==1.55.3
pdf2image==1.17.0
pdfplumber==0.11.4
pydantic==2.10.2
openai==1.55.3
pydub==0.25.1
pypdf==5.1.0
pytesseract==0.3.13
python-dotenv==1.0.1
webdavclient3==3.14.6
Werkzeug==3.1.3
opencv-python==4.10.0.84
